+++ lib/cache/constructor.go
@@ -176,7 +176,7 @@ func (c *Config) UnmarshalYAML(value *yaml.Node) error {
 		aliased.Type = inferredType
 	}
 
-	if spec, exists := pluginSpecs[aliased.Type]; exists {
+	if spec, exists := pluginSpecs[aliased.Type]; exists && spec.confConstructor != nil {
 		confBytes, err := yaml.Marshal(aliased.Plugin)
 		if err != nil {
 			return fmt.Errorf("line %v: %v", value.Line, err)