+++ lib/condition/plugin_test.go
@@ -94,6 +94,9 @@ func TestPluginDescriptions(t *testing.T) {
 			"bar": mConf.Bar,
 		}
 	})
+	RegisterPlugin("foo_no_conf", nil, nil)
+	DocumentPlugin("foo_no_conf", "This is a plugin without config.", nil)
+	RegisterPlugin("foo_no_conf_no_desc", nil, nil)
 
 	exp := `Condition Plugins
 =================
@@ -107,6 +110,8 @@ beyond the standard set.
 
 1. [` + "`bar`" + `](#bar)
 2. [` + "`foo`" + `](#foo)
+3. [` + "`foo_no_conf`" + `](#foo_no_conf)
+4. [` + "`foo_no_conf_no_desc`" + `](#foo_no_conf_no_desc)
 
 ## ` + "`bar`" + `
 
@@ -128,6 +133,12 @@ plugin:
   baz: 10
   foo: default
 ` + "```" + `
+
+## ` + "`foo_no_conf`" + `
+
+This is a plugin without config.
+
+## ` + "`foo_no_conf_no_desc`" + `
 `
 
 	act := PluginDescriptions()